+++ src/domain/raum/index.ts
@@ -1,3 +1,4 @@
 export * from './raum';
+export * from './koordinaten';
 export * from './raum-repository-port';
 export * from './raum-service';